@@ -5,6 +5,7 @@
   "main": "src/index.jsx",
   "private": true,
   "scripts": {
+    "lint-fix": "eslint --fix --ext .js,.jsx .",
     "lint": "eslint --ext .jsx --ext .js src",
     "lint-nice": "eslint --ext .jsx --ext .js src || exit 0;",
     "test": "echo \"Error: no test specified\" && exit 1",
@@ -26,24 +27,18 @@
   "homepage": "https://github.com/OpenSpace/OpenSpace#readme",
   "dependencies": {
     "core-js": "^3.11.2",
-    "d3-scale": "^3.3.0",
     "lodash": "^4.17.21",
     "openspace-api-js": "^0.1.5",
     "prop-types": "^15.7.2",
     "re-resizable": "^6.9.9",
-    "react": "^16.14.0",
     "react-beautiful-dnd": "^13.1.0",
     "react-color": "^2.19.3",
-    "react-dom": "^16.14.0",
     "react-draggable": "^4.4.3",
-    "react-hot-loader": "^4.13.0",
-    "react-input-autosize": "^2.2.2",
-    "react-redux": "^7.2.4",
+    "react-input-autosize": "^3.0.0",
+    "react-redux": "^7.2.8",
+    "react-refresh": "^0.14.0",
     "react-rnd": "^10.3.7",
-    "react-router-dom": "^5.2.0",
-    "react-select": "^3.2.0",
     "react-simple-keyboard": "^3.2.72",
-    "redux": "^4.1.0",
     "sass": "^1.49.9",
     "shallow-equal": "^1.2.1",
     "simple-keyboard-layouts": "^3.0.176",
@@ -51,26 +46,31 @@
   },
   "devDependencies": {
     "@babel/core": "^7.14.0",
+    "@babel/eslint-parser": "^7.18.2",
     "@babel/preset-env": "^7.14.1",
     "@babel/preset-flow": "^7.13.13",
-    "@babel/preset-react": "^7.13.13",
+    "@babel/preset-react": "^7.18.6",
     "@iconify/react": "^3.2.2",
-    "babel-eslint": "^10.1.0",
     "babel-loader": "^8.2.2",
     "copyfiles": "^2.4.1",
-    "cross-env": "^5.2.1",
-    "css-loader": "^3.6.0",
+    "cross-env": "^7.0.3",
+    "css-loader": "^6.7.1",
+    "d3-scale": "^4.0.2",
     "eslint": "^8.10.0",
     "eslint-config-airbnb": "^19.0.4",
     "eslint-plugin-import": "^2.22.1",
     "eslint-plugin-jsx-a11y": "^6.4.1",
     "eslint-plugin-react": "^7.23.2",
-    "file-loader": "^4.3.0",
-    "flow-bin": "^0.101.1",
-    "rimraf": "^2.7.1",
-    "sass-loader": "^7.3.1",
-    "style-loader": "^0.23.1",
-    "url-loader": "^2.3.0",
+    "flow-bin": "^0.182.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^6.3.0",
+    "react-select": "^5.4.0",
+    "redux": "^4.2.0",
+    "rimraf": "^3.0.2",
+    "sass-loader": "^13.0.2",
+    "style-loader": "^3.3.1",
+    "url-loader": "^4.1.1",
     "webpack": "^5.69.1",
     "webpack-cli": "^4.9.2",
     "webpack-dev-server": "^4.7.4"
@@ -86,9 +86,6 @@
       ],
       "@babel/preset-react",
       "@babel/preset-flow"
-    ],
-    "plugins": [
-      "react-hot-loader/babel"
     ]
   }
 }
