@@ -1,8 +1,10 @@
-import React, { Component } from 'react';
+import React from 'react';
 import { connect } from 'react-redux';
-import { withRouter } from 'react-router-dom';
+import { useLocation } from 'react-router-dom';
 import { setShowAbout, startConnection } from '../api/Actions';
-import { formatVersion, isCompatible, RequiredOpenSpaceVersion, RequiredSocketApiVersion } from '../api/Version';
+import {
+  formatVersion, isCompatible, RequiredOpenSpaceVersion, RequiredSocketApiVersion,
+} from '../api/Version';
 import BottomBar from '../components/BottomBar/BottomBar';
 import KeybindingPanel from '../components/BottomBar/KeybindingPanel';
 import environment from '../api/Environment'
@@ -20,128 +22,115 @@ import styles from './OnScreenGui.scss';
 import TourPopup from '../components/GettingStartedTour/TourPopup'
 import { TutorialProvider } from '../components/GettingStartedTour/GettingStartedContext';
 
-class OnScreenGui extends Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-      showTutorial : false,
-      luaConsoleVisible : false
-    }
-    this.checkedVersion = false;
-    this.showFlightController = props.isInBrowser;
-    this.toggleConsole = this.toggleConsole.bind(this);
-  }
+function OnScreenGui({
+  showFlightController, connectionLost, startConnection, version, showAbout, hideAbout, isInBrowser
+}) {
+  let hasCheckedVersion = false;
+  const location = useLocation();
+  const [showTutorial, setShowTutorial] = React.useState(false);
+  const [luaConsoleVisible, setLuaConsoleVisible] = React.useState(false);
 
-  componentDidMount() {
-    this.props.startConnection();
-    window.addEventListener("keydown", this.toggleConsole);
-  }
+  React.useEffect(() => {
+    startConnection();
+    window.addEventListener("keydown", toggleConsole);
+    return () => window.removeEventListener('keydown', toggleConsole);
+  }, []);
 
-  componentWillUnmount() {
-    window.removeEventListener('keydown', this.toggleConsole);
-  }
-
-  toggleConsole(e) {
+ 
+  function toggleConsole(e) {
       if(e.code === "Backquote") {
-        this.setState({
-          luaConsoleVisible : !this.state.luaConsoleVisible
-        });
+        setLuaConsoleVisible(!luaConsoleVisible);
       }
   }
 
-  checkVersion() {
-    if (!this.checkedVersion && this.props.version.isInitialized) {
-      const versionData = this.props.version.data;
-      if (!isCompatible(versionData.openSpaceVersion, RequiredOpenSpaceVersion)) {
-        console.warn(
-          'Possible incompatibility: \nRequired OpenSpace version: ' +
-          formatVersion(RequiredOpenSpaceVersion) +
-          '. Currently controlling OpenSpace version ' +
-          formatVersion(versionData.openSpaceVersion) + '.'
-        );
-      }
-      if (!isCompatible(versionData.socketApiVersion, RequiredSocketApiVersion)) {
-        console.warn(
-          "Possible incompatibility: \nRequired Socket API version: " +
-          formatVersion(RequiredSocketApiVersion) +
-          ". Currently operating over API version " +
-          formatVersion(versionData.socketApiVersion) + '.'
-        );
-      }
-      this.checkedVersion = true;
+  // Check the version
+  if (!hasCheckedVersion && version.isInitialized) {
+    const versionData = version.data;
+    if (!isCompatible(versionData.openSpaceVersion, RequiredOpenSpaceVersion)) {
+      console.warn(
+        `Possible incompatibility: \nRequired OpenSpace version: ${
+          formatVersion(RequiredOpenSpaceVersion)
+        }. Currently controlling OpenSpace version ${
+          formatVersion(versionData.openSpaceVersion)}.`,
+      );
     }
+    if (!isCompatible(versionData.socketApiVersion, RequiredSocketApiVersion)) {
+      console.warn(
+        `Possible incompatibility: \nRequired Socket API version: ${
+          formatVersion(RequiredSocketApiVersion)
+        }. Currently operating over API version ${
+          formatVersion(versionData.socketApiVersion)}.`,
+      );
+    }
+    hasCheckedVersion = true;
   }
 
-  reloadGui() {
+  function reloadGui() {
     location.reload();
   }
 
-  render() {
-    const {isInBrowser} = this.props;
-    this.checkVersion();
-    return (
-      <div className={styles.app} style={environment.developmentMode ? { borderStyle: 'solid', borderWidth: '3px', borderColor: 'orange' } : null}>
-        {environment.developmentMode &&
-          <div className={styles.devModeTextBox}>
-            <p>Dev Gui</p>
-          </div>
-        }
-        <TutorialProvider>
-          {this.props.showAbout && (
+  return (
+    <div className={styles.app} style={environment.developmentMode ? { borderStyle: 'solid', borderWidth: '3px', borderColor: 'orange' } : null}>
+      {environment.developmentMode &&
+        <div className={styles.devModeTextBox}>
+          <p>Dev Gui</p>
+        </div>
+      }
+      <TutorialProvider>
+        { showAbout && (
           <Overlay>
             <Stack style={{ maxWidth: '500px' }}>
-              <Button style={{ alignSelf: 'flex-end', color: 'white' }} onClick={this.props.hideAbout}>
+              <Button style={{ alignSelf: 'flex-end', color: 'white' }} onClick={hideAbout}>
                 Close
               </Button>
               <About />
             </Stack>
           </Overlay>
         )}
-        { this.props.connectionLost && (
+        { connectionLost && (
           <Overlay>
             <Error>
               <h2>Houston, we've had a...</h2>
               <p>...disconnection between the user interface and OpenSpace.</p>
               <p>Trying to reconnect automatically, but you may want to...</p>
-              <Button className={Error.styles.errorButton} onClick={this.reloadGui}>Reload the user interface</Button>
+              <Button className={Error.styles.errorButton} onClick={reloadGui}>Reload the user interface</Button>
             </Error>
           </Overlay>
         )}
         <section className={styles.Grid__Left}>
-          <Sidebar showTutorial={ (show) => this.setState({ showTutorial : show })}/>
+          <Sidebar showTutorial={ setShowTutorial } />
         </section>
         <section className={styles.Grid__Right}>
-          {isInBrowser && this.state.luaConsoleVisible && <LuaConsole />}
+          {isInBrowser && luaConsoleVisible && <LuaConsole />}
           <NodePopOverContainer />
           <NodeMetaContainer />
-          <BottomBar showFlightController={this.showFlightController}/>
+          <BottomBar showFlightController={showFlightController} />
           <KeybindingPanel />
-          <TourPopup isVisible={this.state.showTutorial} setVisibility = { (show) => this.setState({ showTutorial : show })}/>
+          <TourPopup isVisible={showTutorial} setVisibility = { (show) => setShowTutorial(show)}/>
         </section>
       </TutorialProvider>
-      </div>
-    );
-  }
+    </div>
+  );  
 }
 
-const mapStateToProps = state => ({
+const mapStateToProps = (state) => ({
   connectionLost: state.connection.connectionLost,
   version: state.version,
-  showAbout: state.local.showAbout
+  showAbout: state.local.showAbout,
 });
 
-const mapDispatchToProps = dispatch => ({
+const mapDispatchToProps = (dispatch) => ({
   startConnection: () => {
     dispatch(startConnection());
   },
   hideAbout: () => {
-    dispatch(setShowAbout(false))
-  }
+    dispatch(setShowAbout(false));
+  },
 });
 
-OnScreenGui = withRouter(connect(
+OnScreenGui = connect(
   mapStateToProps,
   mapDispatchToProps,
-)(OnScreenGui));
+)(OnScreenGui);
 
 export default OnScreenGui;
